# openapi.yaml
openapi: 3.0.3
info:
  title: iFix RepairTrack API
  version: "1.0.0"
  description: |
    Repair job tracking, warranty management and customer engagement for
    iFix Express.

    Authenticate via POST /api/auth/login, then send the JWT as an
    Authorization Bearer header on every staff route.

    Postman setup notes are in POSTMAN.md alongside this file.
servers:
  - url: https://api.ifixexpress.com.my
    description: Production
  - url: http://127.0.0.1:8787
    description: Local (wrangler dev)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: Invalid credentials }
        code: { type: string, example: UNAUTHORIZED }
        fields:
          type: object
          additionalProperties: { type: string }
...
  responses:
...
    Forbidden:
      description: Wrong branch, or non-admin on an admin route
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Invalid input; see `fields`
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
security:
  - bearerAuth: []
paths:
...
  /api/reviews:
    get:
      tags: [Public]
      summary: List reviews with aggregate stats
      security: []
      parameters:
        - { name: branch_id, in: query, schema: { type: integer } }
        - { name: stars, in: query, schema: { type: integer, minimum: 1, maximum: 5 } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200": { description: Reviews + meta }
    post:
      tags: [Public]
      summary: Submit or edit a review (one per job — upsert)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_id, stars]
              properties:
                job_id: { type: string, example: IFX-00001 }
                stars: { type: integer, minimum: 1, maximum: 5, example: 5 }
                comment: { type: string }
                device_type: { type: string, example: Smartphone }
      responses:
        "201": { description: Created }
        "200": { description: Edited existing review }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/staff:
    get:
      tags: [Staff (admin)]
      summary: List staff (admin only)
      parameters:
        - { name: branch_id, in: query, schema: { type: integer } }
        - { name: role, in: query, schema: { type: string, enum: [admin, technician, staff] } }
        - { name: include_inactive, in: query, schema: { type: boolean } }
      responses:
        "200": { description: Staff list }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Staff (admin)]
      summary: Create a staff account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role, branch_id]
              properties:
                name: { type: string, example: Hafiz }
                email: { type: string, example: hafiz@ifixexpress.com.my }
                password: { type: string, format: password, minLength: 10 }
                role: { type: string, enum: [admin, technician, staff] }
                branch_id: { type: integer, example: 1 }
      responses:
        "201": { description: Created }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Email already in use }
  /api/staff/{id}:
    patch:
      tags: [Staff (admin)]
      summary: Update staff — role, branch, deactivate, or reset password
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                role: { type: string, enum: [admin, technician, staff] }
                branch_id: { type: integer }
                is_active: { type: boolean }
                password: { type: string, format: password, minLength: 10 }
      responses:
        "200": { description: Updated }
        "400": { description: "Validation error, or self-demotion/self-deactivation" }
        "409": { description: Would deactivate the last active admin }